Adding module subdirectories
Added subdirectories for the habitats, metabolisms, and visualizations module files. For now, I've just moved the Mars habitat, methanogens, and Marsvis modules to their appropriate subdirectories. I also edited the QHF.py script to look in these subdirectories for modules, and tested with the Mars config file that it properly finds these modules.